requirements.txt
accelerate==1.2.0
deepspeed==0.15.4
einops==0.8.0
matplotlib==3.10.0
modelscope==1.21.0
numpy==2.2.0

peft==0.14.0
Pillow==11.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
Requests==2.32.3
tiktoken==0.8.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5

transformers==4.47.0
transformers_stream_generator==0.0.5
